Use hooks.
This change replaces the use of `start-process` with the
`make-comint-in-buffer` function, as well as `comint-send-string`
with `comint-simple-send`; thereby lifting the control into comint.

This allows the hooks to work as expected and makes the command output
read-only.

However, due to running all the hooks, this apparently slows down
starting intero a little.