Sort on column as well.
Quickfix already sorts the lines backwards to apply fixes, as they may
change the source location.  This however bearks down if two changes
happen in the same line.  Hence we add the column to the sort criteria
as well, such that corrections in the same line are applied in reverse
as well.