Store temporary working files under .stack-work, not /tmp
See #279. When enabling docker for stack, the docker image cannot see
/tmp paths on the host system. With this change, .stack-work/intero/ is
used as the location for temporary copies of the current working source
file.

This commit does not change the containing directory for *all* intero
temp files, because it is not clearly necessary to do so, but that might
be a logical subsequent step for the sake of consistency.

Moving temp files to .stack-work/intero/ potentially allows intero to
proactively delete all its temp files upon exit, which would be a win,
since the code was previously relying on the OS to tidy up the temp dir.