Customize intero-stack-executable-function instead of a variable
Sometimes, the exact path to the executable might not be known, and
its name not present in user’s PATH. In such a case, the user would
benefit from the possibility of defining their own function to
determine that path programmatically.

One possible ‘improvement’ would be to:

```elisp
(defun intero--stack-executable ()
  (funcall intero-stack-executable-function))
```

… and use that everywhere, instead of `(funcal
intero-stack-executable-function)`. But I didn’t want to introduce
that initially, as another similar symbol might confuse users…