Cache flycheck results when the source is unchanged
This addresses #143, wherein flycheck warnings were getting lost after
two flycheck passes in quick succession since the second underlying :l
command simply said "ok".

The approach taken here is to note the temp buffer modification time
and results for a flycheck pass, and then to simply replay them if
another pass is requested with the same temp buffer mod time.

In order to make this work, we have to ensure that the mod time of the
temp buffer is not updated unless its contents change, which
introduces a small inefficiency: the temp file is now first quickly
read before conditionally writing to it.