Change completeIdentifier to how GHCi does it
Fixes issue where `:complete repl "<$"` lists everything in scope
instead of completing the symbolic identifier.

Source for completeIdentifier comes from:
https://github.com/ghc/ghc/blob/b8fec6950ad99cbf11cd22698b8d5ab35afb828f/ghc/GHCi/UI.hs#L3021

As of base-4.10, Text.Read.Lex exports isSymbolChar. I didn't want to
mess with the dependencies, so I found the source at this link, and
used the imports from Data.Char to re-write it in source.
http://hackage.haskell.org/package/base-4.10.0.0/docs/src/Text.Read.Lex.html#isSymbolChar