Tab always completes in repl, compnay-idle never pops up anymore
Tab should always trigger autocopmlete in the repl. This makes the
intero repl mode more like an actual ghci. I see very little use in
tab indentation in the repl. However, I realize I may be pushing this
opinion on anyone who does, so I am not opposed to removing this.

company-idle-delay is set to 0.5 in prelude (which I use). The
suggestions are not using repl-tab complete, so they are not
helpful.