Document that `C-c C-k` quits the suggestion pane
We used to be able to quit the suggestion pane with `C-g`, but that
was removed in 397e5c8258c97be8e2abb1d34526219fc816ca86. The the `C-g`
was intuitive to me, but `C-c C-k` was not -- I had to look at the Git
log to figure it out -- so I thought it would be helpful to document
it. My next guess after `C-g` was `q`, which is also more intuitive
than `C-c C-k` to me.

I did not test this change, since I don't know how to run Emacs
packages from source.