More robustly delete network process buffers.
See #544. Unfortunately I seem to have the worst of both worlds. With
2c2546b I get "selecting deleted buffer" errors, sometimes causing emacs
to hang briefly; and with 37af015 my buffer list continues to grow,
though maybe slower than before. (I'm using emacs 25.1.1.)

It appears that for me, the "deleted\n" event is only getting triggered
inconsistently when `(delete-process)` is called. So I return to killing
the buffer immediately. But I also handle the deleted event to avoid
errors.

I considered having

    (when (buffer-live-p (process-buffer process))
      (kill-buffer (process-buffer process)))

inside the deleted event, as a fallback. But it doesn't seem to get
called for me in normal workflow, so if something unusual is happening,
I decided to not destroy the evidence of it.