Use `fno-code fwrite-interface` when GHC >= 8.4.1
Before this commit, we pass `-fobject-code` to ghci, which produces
build artifacts (object files). This allows subsequent reloads to be
faster and improves the performance of intero.

It is possible to output interface files while only doing the work of
typechecking with the flags `fno-code` and `fwrite-interface`. This,
again, improves the performance of intero.

Unfortunately, `fno-code` does not work well with Template Haskell
before GHC version 8.4.1. Therefore, we switch which flags we provide
based on version number.